+++ my-app/src/components/Banner.js
@@ -1,0 +1,9 @@
+export const Banner = () => {
+  return (
+    <div className="trail text-center mb-4">
+      <p className="mb-0">
+        <strong>Try it free 7 days</strong> then $20/mo. thereafter
+      </p>
+    </div>
+  );
+};
